Working packet sniffer
Returns a max number of connections, with this statistics. Collects
information per connection such as src and dst ip & port, protocol
type, packet count.

This could be extended by a lot. And is actually a good project to keep
working on. I think it'd be useful to continue working on this.